+++ src/ewokscore/tests/serialization/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest.register_assert_rewrite("ewokscore.tests.utils.serialization.example_data")